FlakeFinder/requirements.txt
numpy
matplotlib
opencv-python
scikit-image
pywin32
mysql-connector-python